pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "diarization-benchmarks"
version = "0.1.0"
description = "CPU diarization benchmark service with FastAPI and Fly.io deployment"
readme = "README.md"
license = "CC-BY-4.0"
requires-python = ">=3.10"
authors = [
  {name = "Alexander Lukashov"}
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "foxnose-sdk==0.4.2",
  "pyannote.core>=5.0.0",
  "pyannote.metrics>=3.2.1",
  "soundfile>=0.12.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
  "app",
  "config",
  "scripts",
  "pyproject.toml",
  "README.md"
]
